+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### 0.22.8 - 2020-09-08
+### UNRELEASED
 
  - Fix aioredis data source builder