+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### UNRELEASED
+### 0.22.10 - 2020-09-09
 
  - Add optional support for newrelic on aioredis data source